requirements.txt
Flask==3.0.3
Flask-PyMongo==3.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
dnspython==2.6.1
